fixed apply for reivew issue

+++ app/Data/Repositories/JobBidRepository.php
@@ -340,11 +340,14 @@ public function update(array $data = [])
 
         }
         $data = parent::update($data);
-            
-        $criteria = ['job_id' , $data->job_id, 'is_visit_required' => 1];
+
+        $criteria = ['job_id' => $data->job_id, 'is_visit_required' => 1];
 
         if($status == 'awarded'){
-            $this->model->where($criteria)->delete();
+
+            if (!\App::runningInConsole()) {
+                $this->model->where($criteria)->delete();
+            }        
         }
 
     }
